MASSIVE simplification of the ray-cone intersection
It was, admittedly, a huge pile of !@#$%%# ...

Completely overhauled, and now also uses a more physically correct
scattering computation (using the Beer-Lambert law).  Should now also
be a fair bit faster, thanks to better utilizations of assumptions.

Unfortunately, the issue, that caused this excursion in the first
place, is still there. Namely, that the "bottom end" of the cone is
strangely, and homogenously bright.