Simplified SSBO structure for clustered shading
No more "global index lists" w/ ranges and stuff :)
Everything cluster-related is contained in the cluster definition
itself.
Might not be optimal in terms of performance (no idea), but it is a hell of a
lot easier to understand. It does take a tool on shader link
time. (large(r) structs does that) :(

Also introduced a code generation script to the build pipeline. The
struct definitions in "shared-structs.glh" generate the C++
counterparts (with alignment and padding).

Also, this introduces a persistently-mapped SSBO containing the
lights (in a LightsManagement struct).  Again, it might not be optimal
performance, but it does - or I hope it does - make the light updates
and meta data, especially for the forthcoming shadow map
implementation.

This also marks a beginning of the shadow map implementation, which
will use a (single) large texture atlas. Currently it generates shadow
maps for the first point light in the scene. The atlas can be viewed
in the render target debug visualization.