Basic shadow mapping successful!
It's pretty hard-coded now for only a single point light, but it
atleast uses a shadow atlas (4k x 4k, for now).

A new SSBO was added to associate the shadow mapping parameters, like
the atlas rectangle (the view projections, for points 6, one for each
face).
There's a bunch of bias tweking parameters, but in the current test
room, none of them are needed. There is a slight shadow detachment,
however.

It does 3x3 gaussian sampling (PCF?), but more blurring is required.

TODO: is more lights, to see how many can be done without tanking performance.