Use UBO for light counts
Will make it even better when we'll merge all the light types into a
single struct, so the SSBO can be a dynamic-sized array (instead of a
literal fixed-sized array).

Also, lowering local_size_x for cull_light compute shader (1024 ->
256) is A LOT faster. Who knew... ;)
Have more ideas for improvement, but so far just added as a comment.