Renamed "radius" -> "affect_radius"
Clarifying it means the radius which the light affects the
environment.
This also implies that it needs to be computed CPU-side whenever the
light intensity cchanges.
Makes light culling a bit simpler; the only exception is spot lights,
which use a different sphere (which also needs to be computed
CPU-side, whenever intensity or outer cone angle changes.

Also added stubs for all light types in the shader code.