WIP Shadow atlas slot allocation logic shaping up
But it seems complicated... At least time-wise, it needs to be
smarter. Since from one frame to the next, it's very likely there will
be no change. So that code path needs to be quick.

However, the shadow slot thing could probably run in parllel to rendering, at a fixed interval
, say ~5Hz.
Currently, on my meager i5 8400 it takes about 3ms, which is quite
long.
I suppose it could be done in a compute shader; light value
computation all in parallel, and then slot allocation in some "clever"
way ;)