Added size-based rendering interval for shadow maps
Smaller slots are rendered less often.
This is currently measured in absolute time (ms), but maybe it should
instead be number of frames (or scale by current average FPS).
That way, it doesn't completely tank the system if FPS is low.
This is better than nothing, however.
The render intervals are currently hard-coded, should be configurable, obviously.