There's a slight chance that the dynamic shadow atlas now works!
It switches sizes dynamically depending on light value (importance),
and it actually looks really good. It's very difficult to see the
switch as well. :)

Moved the cube face deduction to a separate function.
The *LightVisibility() functions now return vec3, so in theory a
shadow by a color filter. Not sure if that's really necessary
though... ;)