Made the atlas log for concise
The various counters is now only output if non-zero.
Also the size+count listing it now printed using symols/icons instead
f text, and also on the same line (to total one line is printed).

Also added a simple assert that allocated+free slots is equal to the total
number of slots.