Don't render shadow maps for lights that didn't contribute to the frame
cull_lights shader writes into a bitmask for each light it
collected (currently max 1024 bits/lights = 32 uints).
That bitmask SSBO is downloaded before each frame (thus the state from
previous frame); simple hack to avoid stalls.
When rendering shadow maps, all lights not contained in that set of
lights is simly skipped; it won't show up on screen any way.

With this, the frustum-sphere culling of lights is not necessary any
more. I guess it's still needed to limit the number of lights that is
input to the cull_lights shader, e.g. only consider those that are
"sufficiently close".