Calculate near plane for shadow map rendering a bit more dynamically
near_z = max(0.1f, far_z / 250.f)

Ideally, the near plane should be as far away as possible, but closer than the
closest shadow caster. The far plane, conversely, should be as close
as possible, but further away than the furthest shadow receiver. That
is the shortest range that still encompasses the objects involved.
But since that's a bit difficult to discern, let's go with this
compromise, which shouldn't be too outrageous.