Added support for tube lights
Not very clever; it simlulates it was a thin area light always facing
the fragment.

Also volumetrics supported, even less clever, just computes the
nearest distance to the tube light and use the regular falloff with that.

The affect_radius needs to be toned down; inherited "madness" from
area light.